+++ src/v5_hal/firmware/include/main.h
@@ -47,7 +47,7 @@
 #include "DataNodes/ProsTimeNode.h"
 #include "DataNodes/BatteryNode.h"
 #include "DataNodes/InertialSensorNode.h"
-#include "DriverControlNode.h"
+#include "DataNodes/DriverControlNode.h"
 
    /**
     * You should add more #includes here
