+++ src/v5_hal/firmware/src/DataNodes/ADIDigitalInNode.cpp
@@ -30,8 +30,6 @@ int ADIDigitalInNode::getValue() {
 void ADIDigitalInNode::periodic() {
     // Publish data when called, and spin the handler to send data to the
     // coprocessor on the published topic
-    m_populateMessage();
-    // m_publisher->publish(&m_digital_in_msg);
 }
 
 void ADIDigitalInNode::m_populateMessage() {
