+++ src/v5_hal/firmware/src/DataNodes/DriverControlNode.cpp
@@ -1,4 +1,4 @@
-#include "DriverControlNode.h"
+#include "DataNodes/DriverControlNode.h"
 
 // By default, this constructor calls the constructor for the Node object in
 // NodeManager.h
