+++ src/v5_hal/firmware/src/NodeManager.cpp
@@ -14,6 +14,8 @@ ros::NodeHandle* NodeManager::addNode(Node* node,
 }
 
 void NodeManager::initialize() {
+    m_handle->initNode();
+
     for (auto node_structure : m_node_structures) {
         node_structure.node->initialize();
     }
@@ -30,6 +32,7 @@ void NodeManager::execute() {
             node_structure.last_executed_millis = current_time;
         }
     }
+    
     pros::c::delay(m_delay_time_millis);
 }
